Now use fixtures to teardown both test_multiple_pvc_creation_deletion_scale and test_all_4_type_pvc_creation_deletion_scale methods.
Refactored the collection of csi-provisioner logs used for determining
the time spent running.  In doing so, added a check to insure that
IndexErrors do not occur if csi-provisioner data is not found
(workaround until Bugzilla 1812798 is fixed).

Made sure deletes of individual object lists do not wait for each
individual object.  Instead, after deletes are performed, the list
is rechecked and waits are performed as needed.

Cleaned up erroneous code found in previous reviews.

Signed-off-by: wusui <wusui@redhat.com>